relocated PluginFeature
